chore(dev): move generate demo stories script to dev-packages
